backend/pyproject.toml
[project]
name = "excel-ai-agent-backend"
version = "0.1.0"
description = "Secure FastAPI backend for Excel AI Agent"
authors = [
    {name = "Brayan Pineda",email = "pinedab606@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13, <4.0"


dependencies = [
    #Core FastAPI
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0", #It takes HTTP requests from the outside world → passes them to your FastAPI app → returns responses.
    # Authentication & Security
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    #Configuration Env
    "pydantic[email]>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.0",
    # Database
    "supabase>=2.18.1,<3.0.0",
    #AI Model
    "google-genai (>=1.36.0,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "formulas (>=1.3.1,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "docker (>=7.1.0,<8.0.0)",
    "pandas (>=2.3.3,<3.0.0)"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "ruff>=0.1.6",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
name = "excel-ai-agent-backend"
version = "0.1.0"
description = "Secure FastAPI backend for Excel AI Agent"
authors = ["Brayan Pineda <pinedab606@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
pytest-asyncio = "^1.2.0"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py313"
line-length = 88
select = ["E", "F", "W", "C90", "I", "N", "UP", "YTT", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "DJ", "EM", "EXE", "FA",
"ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SLOT", "SIM", "TID", "TCH", "INT", "ARG", "PTH", "ERA",
"PD", "PGH", "PL", "TRY", "FLY", "NPY", "AIR", "PERF", "FURB", "LOG", "RUF"]

